# Optimize writer