Some optimizations for #250

```
benchmark                                      old ns/op     new ns/op     delta
BenchmarkEJ_Unmarshal_M-16                     33735         33282         -1.34%
BenchmarkEJ_Unmarshal_S-16                     457           456           -0.22%
BenchmarkEJ_Marshal_M-16                       12328         10483         -14.97%
BenchmarkEJ_Marshal_L-16                       594185        526980        -11.31%
BenchmarkEJ_Marshal_L_ToWriter-16              511792        417277        -18.47%
BenchmarkEJ_Marshal_M_Parallel-16              3288          2908          -11.56%
BenchmarkEJ_Marshal_M_ToWriter-16              11010         9164          -16.77%
BenchmarkEJ_Marshal_M_ToWriter_Parallel-16     1675          1363          -18.63%
BenchmarkEJ_Marshal_L_Parallel-16              131516        119829        -8.89%
BenchmarkEJ_Marshal_L_ToWriter_Parallel-16     88025         69712         -20.80%
BenchmarkEJ_Marshal_S-16                       162           159           -1.85%
BenchmarkEJ_Marshal_S_Parallel-16              45.7          39.2          -14.22%

benchmark                                      old MB/s     new MB/s     speedup
BenchmarkEJ_Unmarshal_M-16                     386.06       391.32       1.01x
BenchmarkEJ_Unmarshal_S-16                     179.14       179.70       1.00x
BenchmarkEJ_Marshal_M-16                       725.72       853.42       1.18x
BenchmarkEJ_Marshal_L-16                       752.98       849.01       1.13x
BenchmarkEJ_Marshal_L_ToWriter-16              874.20       1072.21      1.23x
BenchmarkEJ_Marshal_M_Parallel-16              3960.40      4477.48      1.13x
BenchmarkEJ_Marshal_M_ToWriter-16              812.57       976.31       1.20x
BenchmarkEJ_Marshal_M_ToWriter_Parallel-16     5338.37      6560.16      1.23x
BenchmarkEJ_Marshal_L_Parallel-16              3401.94      3733.72      1.10x
BenchmarkEJ_Marshal_L_ToWriter_Parallel-16     5082.72      6417.97      1.26x
BenchmarkEJ_Marshal_S-16                       498.12       507.56       1.02x
BenchmarkEJ_Marshal_S_Parallel-16              1773.40      2067.52      1.17x

benchmark                                      old allocs     new allocs     delta
BenchmarkEJ_Unmarshal_M-16                     128            128            +0.00%
BenchmarkEJ_Unmarshal_S-16                     3              3              +0.00%
BenchmarkEJ_Marshal_M-16                       10             10             +0.00%
BenchmarkEJ_Marshal_L-16                       29             29             +0.00%
BenchmarkEJ_Marshal_L_ToWriter-16              24             24             +0.00%
BenchmarkEJ_Marshal_M_Parallel-16              9              9              +0.00%
BenchmarkEJ_Marshal_M_ToWriter-16              8              8              +0.00%
BenchmarkEJ_Marshal_M_ToWriter_Parallel-16     8              8              +0.00%
BenchmarkEJ_Marshal_L_Parallel-16              33             33             +0.00%
BenchmarkEJ_Marshal_L_ToWriter_Parallel-16     24             24             +0.00%
BenchmarkEJ_Marshal_S-16                       1              1              +0.00%
BenchmarkEJ_Marshal_S_Parallel-16              1              1              +0.00%

benchmark                                      old bytes     new bytes     delta
BenchmarkEJ_Unmarshal_M-16                     9792          9792          +0.00%
BenchmarkEJ_Unmarshal_S-16                     128           128           +0.00%
BenchmarkEJ_Marshal_M-16                       10378         10381         +0.03%
BenchmarkEJ_Marshal_L-16                       464054        463670        -0.08%
BenchmarkEJ_Marshal_L_ToWriter-16              2623          2596          -1.03%
BenchmarkEJ_Marshal_M_Parallel-16              10702         10790         +0.82%
BenchmarkEJ_Marshal_M_ToWriter-16              741           741           +0.00%
BenchmarkEJ_Marshal_M_ToWriter_Parallel-16     773           777           +0.52%
BenchmarkEJ_Marshal_L_Parallel-16              520109        525033        +0.95%
BenchmarkEJ_Marshal_L_ToWriter_Parallel-16     4191          4042          -3.56%
BenchmarkEJ_Marshal_S-16                       128           128           +0.00%
BenchmarkEJ_Marshal_S_Parallel-16              128           128           +0.00%

```